+++ init.lua
@@ -1,7 +1,7 @@
 vim.cmd('ru! innit.vim')
 
+require('o')
 require('colo')
-require('core')
 require('plugins')
 
 if vim.g.neovide then require('neovide') end
