--- lua/core/init.lua
@@ -1,5 +1,0 @@
-require('core.o')
-require('core.keymaps').core()
-
-vim.g.mapleader = ' '
-vim.g.maplocalleader = ' '
