+++ lua/plugins/behavior/cmp.lua
@@ -2,8 +2,6 @@ local cmp = require('cmp')
 local luasnip = require('luasnip')
 local cmp_autopairs = require('nvim-autopairs.completion.cmp')
 
-local keymaps = require('core.keymaps').cmp(cmp, luasnip)
-
 cmp.event:on(
   'confirm_done',
   cmp_autopairs.on_confirm_done()
@@ -20,17 +18,19 @@ local function notcomment()
     return not treesitter and not syntax
 end
 
-cmp.setup({
-    enabled = notcomment,
-    snippet = {
-        expand = function(args) luasnip.lsp_expand(args.body) end,
-    },
-    mapping = cmp.mapping.preset.insert(keymaps),
-    sources = cmp.config.sources({
-        { name = 'nvim_lua' }, -- one day I will develop my nvim config in nvim
-        { name = 'nvim_lsp' },
-        { name = 'luasnip' },
-        { name = 'buffer' },
-        { name = 'nvim_lsp_signature_help' },
+return function(keymaps)
+    cmp.setup({
+        enabled = notcomment,
+        snippet = {
+            expand = function(args) luasnip.lsp_expand(args.body) end,
+        },
+        mapping = cmp.mapping.preset.insert(keymaps(cmp, luasnip)),
+        sources = cmp.config.sources({
+            { name = 'nvim_lua' }, -- one day I will develop my nvim config in nvim
+            { name = 'nvim_lsp' },
+            { name = 'luasnip' },
+            { name = 'buffer' },
+            { name = 'nvim_lsp_signature_help' },
+        })
     })
-})
+end
