+++ lua/plugins/behavior/comment.lua
@@ -1,6 +1,6 @@
-local keymaps = require('core.keymaps').comment()
-
-require('Comment').setup(vim.tbl_extend('force', {
+return function(keymaps)
+    require('Comment').setup(vim.tbl_extend('force', {
     sticky = false,
     ignore = '^$', -- ignore empty lines
-}, keymaps))
+    }, keymaps))
+end
