+++ lua/plugins/init.lua
@@ -38,14 +38,17 @@ require('packer').startup(function(use)
     if packer_bootstrap then require('packer').sync() end
 end)
 
+-- prepare keymaps first
+local k = require('plugins.whichkey')
+
 -- plugin setups, sorted alphabetically
-require('plugins.external.lsp')
+require('plugins.external.lsp')(k.on_attach)
 require('plugins.external.null')
-require('plugins.external.dap')
+require('plugins.external.dap')(k.dap)
 
 require('plugins.behavior.autopairs')
-require('plugins.behavior.cmp')
-require('plugins.behavior.comment')
+require('plugins.behavior.cmp')(k.cmp)
+require('plugins.behavior.comment')(k.comment)
 require('plugins.behavior.luasnip')
 require('plugins.behavior.surround')
 require('plugins.behavior.telescope')
@@ -53,6 +56,5 @@ require('plugins.behavior.telescope')
 require('plugins.interface.bufferline')
 require('plugins.interface.indentline')
 require('plugins.interface.lualine')
-require('plugins.interface.nvim-tree')
+require('plugins.interface.nvim-tree')(k.nvimtree)
 require('plugins.interface.treesitter')
-require('plugins.interface.whichkey')
