+++ lua/plugins/interface/nvim-tree.lua
@@ -1,21 +1,26 @@
-require('nvim-tree').setup({
-    disable_netrw = true,
-    open_on_setup = true,
-    open_on_tab = true,
-    hijack_cursor = true,
-    diagnostics = {
-        enable = true,
-        show_on_dirs = true,
-    },
-    view = {
+local tree =  require('nvim-tree')
+local api = require('nvim-tree.api')
 
-    },
-    renderer = {
-        add_trailing = true,
-        highlight_git = true,
-        icons = {
-            git_placement = 'signcolumn',
-        }
-    },
-})
-require('core.keymaps').nvimtree()
+return function(keymaps)
+    tree.setup({
+        disable_netrw = true,
+        open_on_setup = true,
+        open_on_tab = true,
+        hijack_cursor = true,
+        diagnostics = {
+            enable = true,
+            show_on_dirs = true,
+        },
+        view = {
+
+        },
+        renderer = {
+            add_trailing = true,
+            highlight_git = true,
+            icons = {
+                git_placement = 'signcolumn',
+            }
+        },
+    })
+    keymaps(api)
+end
